requirements.txt
adagio==0.2.6
altair==5.4.1
appdirs==1.4.4
atpublic==5.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
datacompy==0.13.2
fs==2.4.16
fsspec==2024.6.1
fugue==0.9.1
gitdb==4.0.11
GitPython==3.1.43
ibis-framework==9.3.0
idna==3.8
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.6.0
numpy==1.26.4
oracledb==2.4.1
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
parsy==2.1
pillow==10.4.0
polars==0.20.31
protobuf==5.28.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
sample-data==0.1.0
setuptools==74.0.0
six==1.16.0
smmap==5.0.1
sqlglot==25.9.0
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
triad==0.9.8
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid7==0.1.0
watchdog==4.0.2
